chore(kernels): split the `remeshing` module into submodules
